Separate publisher implementations
